refactor main and load_env plus just print for now
We can't use tweepy.API, because the twitter bot
account is only allowed to use twitter API v2,
which is compatible with tweepy.Client().
Also the loadenv file needed to be refactored
because it wasn't working.
Also changed the logic in main.py to suit the
new reminders.txt format.
Also created an example .env-example file to
show the shape of what the .env files needs
to be, and added .env to .gitignore.